requirements.txt
matplotlib
numpy
opencv-python
pillow
pypng
pyyaml
scipy
cython
prettytable
tensorflow-gpu
tensorflow-probability
